Consolidated remaining ledger functionality
