--- src/main/java/cashpiles/model/ModelItem.java
@@ -1,14 +1,0 @@
-package cashpiles.model;
-
-class ModelItem implements Cloneable {
-
-	@Override
-	public ModelItem clone() {
-		try {
-			return (ModelItem) super.clone();
-		} catch (CloneNotSupportedException ex) {
-			throw new RuntimeException(ex);
-		}
-	}
-
-}
